+++ lib/leiratech_trackzero.dart
@@ -7,5 +7,4 @@ part 'network_layer.dart';
 part 'trackzero_api.dart';
 part 'trackzero_client.dart';
 part 'trackzero_entity.dart';
-part 'trackzero_event.dart';
 part 'trackzero_response.dart';