+++ lib/network_layer.dart
@@ -2,11 +2,14 @@ part of leiratech_trackzero;
 
 final _baseUrl = "https://api.trackzero.io";
 
-_request(String endpointUrl, body, String method) async {
+_request(String endpointUrl, body, String method, [headers]) async {
   try {
     var url = Uri.parse('$_baseUrl$endpointUrl');
     var response;
     switch (method) {
+      case "GET":
+        response = await http.get(url, headers: headers);
+        break;
       case "POST":
         response = await http.post(url, body: body);
         break;
@@ -20,6 +23,10 @@ _request(String endpointUrl, body, String method) async {
   } catch (e) {
     return new Response(null, null, e.toString());
   }
+}
+
+get(String endpointUrl, headers) async {
+  return await _request(endpointUrl, {}, "GET", headers);
 }
 
 post(String endpointUrl, body) async {