+++ lib/trackzero_api.dart
@@ -12,25 +12,17 @@ class LogApi {
   static Future<Response> deleteEntity(apiKey, body) async {
     return await delete('$endPoint/entities?X-API-KEY=$apiKey', body);
   }
-
-  ///Creates/Updates the Event
-  static Future<Response> upsertEvent(apiKey, body) async {
-    return await post('$endPoint/events?X-API-KEY=$apiKey', body);
-  }
-
-  ///Deletes the Event
-  static Future<Response> deleteEvent(apiKey, body) async {
-    return await delete('$endPoint/events?X-API-KEY=$apiKey', body);
-  }
 }
 
-class ConfigApi {
-  static const endPoint = "/dynamicconfiguration";
+class SpaceApi {
+  static const endPoint = "/AnalyticsSpaces";
 
-  ///Queries the configuration
-  static Future<Response> queryConfiguration(apiKey, groupId, body) async {
-    return await post(
-        '$endPoint/applicable?X-API-KEY=$apiKey&configurationGroupId=$groupId',
-        body);
+  ///Gets a space portal session
+  static Future<Response> getSession(
+      String apiKey, int analyticsSpaceId, int ttl) async {
+    return await get(
+        '$endPoint/session?analyticsSpaceId=$analyticsSpaceId&ttl=$ttl', {
+      "X-API-KEY": apiKey,
+    });
   }
 }